chore(KAN-251): update imports in all moved files, replacement of relative paths with aliases
