+++ src/app/components/Atoms/Divider/divider.component.scss
@@ -1,4 +1,4 @@
-@use '../../../styles/variables' as *;
+@use 'variables' as *;
 
 .divider {
   &--horizontal {
