+++ src/app/components/Atoms/Icon/icon.component.scss
@@ -1,4 +1,4 @@
-@use '../../../styles/variables' as *;
+@use 'variables' as *;
 
 :host {
   vertical-align: middle;
