+++ src/app/components/Atoms/Link/link/link.component.scss
@@ -1,4 +1,4 @@
-@use '../../../../styles/variables' as *;
+@use 'variables' as *;
 
 .link {
   display: flex;
