+++ src/app/components/Atoms/Toggle-button/toggle-btn.component.scss
@@ -1,4 +1,4 @@
-@use '../../../styles/variables' as *;
+@use 'variables' as *;
 
 .toggle-btn__wrapper {
   display: inline-flex;
