+++ src/app/components/Molecules/Card-ready-to-use/big-visual-trip-card/big-visual-trip-card.component.ts
@@ -1,10 +1,10 @@
 import { Component, Input } from '@angular/core';
-import { CardBodyComponent } from '../../card/card-body/card-body.component';
-import { CardFooterComponent } from '../../card/card-footer/card-footer.component';
-import { CardHeaderComponent } from '../../card/card-header/card-header.component';
-import { CardComponent } from '../../card/card.component';
-import { VariantTripCard } from '../../../model/visual-trip-card';
-import { ChipComponent } from '../../Atoms/chip/chip.component';
+import { VariantTripCard } from '@model/visual-trip-card.model';
+import { CardBodyComponent } from 'components/Atoms/Card/card-body/card-body.component';
+import { CardFooterComponent } from 'components/Atoms/Card/card-footer/card-footer.component';
+import { CardHeaderComponent } from 'components/Atoms/Card/card-header/card-header.component';
+import { CardComponent } from 'components/Atoms/Card/card.component';
+import { ChipComponent } from 'components/Atoms/chip/chip.component';
 
 @Component({
   selector: 'app-big-visual-trip-card',
