+++ tsconfig.json
@@ -19,7 +19,13 @@
     "importHelpers": true,
     "target": "ES2022",
     "module": "ES2022",
-    "lib": ["ES2022", "dom"]
+    "lib": ["ES2022", "dom"],
+    "baseUrl": "src/app",
+    "paths": {
+      "@model/*": ["shared/model/*"],
+      "@dto/*": ["shared/dto/*"],
+      "@service/*": ["core/services/*"]
+    }
   },
   "angularCompilerOptions": {
     "enableI18nLegacyMessageIdFormat": false,
